requirements.txt
fastapi
uvicorn
numpy
tensorflow
requests
Pillow